requirements.txt
fastapi==0.114.1
numpy==1.24.0
mangum==0.17.0
pillow==10.4.0
python-multipart==0.0.9
uvicorn==0.30.6